Music Working, Plays while game is running, and Loops
