+++ TKGame/Art.cs
@@ -21,11 +21,11 @@ public static Texture2D BackgroundTexture { get; private set; }
         /// <param name="content"></param>
         public static void LoadContent(ContentManager content)
         {
-            PlayerTexture = content.Load<Texture2D>(@"Art/Player"); // Will need to be modified when we create a better structure for our files.
-            WeaponTexture = content.Load<Texture2D>(@"Art/Weapons/IronSword"); //Sword Weapon Image    
-            BackgroundTexture = content.Load<Texture2D>(@"Art/Cobble"); //Background image
-            EnemyTexture = content.Load<Texture2D>(@"Art/DoomguyRightFacing"); // enemy as doomguy
-            ItemTexture = content.Load<Texture2D>(@"Art/DiamondSwordLeftTilt"); // item as diamond sword 
+            PlayerTexture = content.Load<Texture2D>(@"C:/Users/Greywater/Documents/Git/TKGame/TKGame/Content/bin/DesktopGL/Art/Player"); // Will need to be modified when we create a better structure for our files.
+            WeaponTexture = content.Load<Texture2D>(@"C:/Users/Greywater/Documents/Git/TKGame/TKGame/Content/bin/DesktopGL/Art/Weapons/IronSword"); //Sword Weapon Image    
+            BackgroundTexture = content.Load<Texture2D>(@"C:/Users/Greywater/Documents/Git/TKGame/TKGame/Content/bin/DesktopGL/Art/Cobble"); //Background image
+            EnemyTexture = content.Load<Texture2D>(@"C:/Users/Greywater/Documents/Git/TKGame/TKGame/Content/bin/DesktopGL/Art/DoomguyRightFacing"); // enemy as doomguy
+            ItemTexture = content.Load<Texture2D>(@"C:/Users/Greywater/Documents/Git/TKGame/TKGame/Content/bin/DesktopGL/Art/DiamondSwordLeftTilt"); // item as diamond sword 
         }
     }
 }
