+++ TKGame/Content/Content.mgcb
@@ -49,32 +49,6 @@
 /processorParam:TextureFormat=Color
 /build:Art/DoomguyRightFacing.png
 
-#begin Art/DoomguyRightFacing.png
-#begin Art/Cobble.jpg
-/importer:TextureImporter
-/processor:TextureProcessor
-/processorParam:ColorKeyColor=255,0,255,255
-/processorParam:ColorKeyEnabled=True
-/processorParam:GenerateMipmaps=False
-/processorParam:PremultiplyAlpha=True
-/processorParam:ResizeToPowerOfTwo=False
-/processorParam:MakeSquare=False
-/processorParam:TextureFormat=Color
-/build:Art/DoomguyRightFacing.png
-/build:Art/Cobble.jpg
-
-#begin Art/Player.png
-/importer:TextureImporter
-/processor:TextureProcessor
-/processorParam:ColorKeyColor=255,0,255,255
-/processorParam:ColorKeyEnabled=True
-/processorParam:GenerateMipmaps=False
-/processorParam:PremultiplyAlpha=True
-/processorParam:ResizeToPowerOfTwo=False
-/processorParam:MakeSquare=False
-/processorParam:TextureFormat=Color
-/build:Art/Player.png
-
 #begin Art/Player.png
 /importer:TextureImporter
 /processor:TextureProcessor
@@ -98,4 +72,10 @@
 /processorParam:MakeSquare=False
 /processorParam:TextureFormat=Color
 /build:Art/Weapons/IronSword.png
+
+#begin Audio/Break-Out-Loop.mp3
+/importer:Mp3Importer
+/processor:SongProcessor
+/processorParam:Quality=Best
+/build:Audio/Break-Out-Loop.mp3
